Refactor: Input 컴포넌트 저장 위치 변경
