--- src/components/common/Input.module.scss
@@ -1,39 +1,0 @@
-@import '/src/styles/constants/color';
-@import '/src/styles/mixins/font';
-
-.inputContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 0.8rem;
-}
-
-.inputTitle {
-  @include text-16-l;
-
-  line-height: 26px; /* 162.5% */
-  color: $black;
-}
-
-.inputBox {
-  width: 35rem;
-  padding: 1.6rem 2rem;
-  font-size: 16px;
-  color: $black;
-  background: $white;
-  border: 0.1rem solid #cbc9cf;
-  border-radius: 6px;
-}
-
-.inputBoxError {
-  width: 35rem;
-  padding: 1.6rem 2rem;
-  font-size: 16px;
-  color: $black;
-  background: $white;
-  border: 0.2rem solid $red-40;
-  border-radius: 0.6rem;
-}
-
-.error {
-  color: $red-40;
-}
